+++ src/PLegendre.f95
@@ -40,12 +40,12 @@ subroutine PLegendre(p, lmax, z, exitstatus)
 
     implicit none
 
-    integer, intent(in) :: lmax
+    integer(int32), intent(in) :: lmax
     real(dp), intent(out) :: p(:)
     real(dp), intent(in) :: z
-    integer, intent(out), optional :: exitstatus
+    integer(int32), intent(out), optional :: exitstatus
     real(dp) :: pm2, pm1, pl
-    integer :: l
+    integer(int32) :: l
 
     if (present(exitstatus)) exitstatus = 0
