+++ src/PlON.f95
@@ -39,12 +39,12 @@ subroutine PlON(p, lmax, z, exitstatus)
 
     implicit none
 
-    integer, intent(in) :: lmax
+    integer(int32), intent(in) :: lmax
     real(dp), intent(out) :: p(:)
     real(dp), intent(in) :: z
-    integer, intent(out), optional :: exitstatus
+    integer(int32), intent(out), optional :: exitstatus
     real(dp) :: pm2, pm1, pl, pi
-    integer :: l
+    integer(int32) :: l
 
     if (present(exitstatus)) exitstatus = 0
